# indexer/api/openapi.yaml
openapi: 3.1.0
info:
  title: Game Components Indexer API
  version: 1.0.0
  description: |
    REST API for querying indexed game token data from the Game Components protocol.

    This API provides access to:
    - Token details with decoded packed metadata
    - Game leaderboards with rankings
    - Player portfolios and activity
    - Game statistics and analytics

    All token IDs are represented as strings due to their large size (felt252/u256).
  contact:
    name: Game Components Team
    url: https://github.com/Provable-Games/game-components
  license:
    name: BUSL-1.1
    url: https://github.com/Provable-Games/game-components/blob/main/LICENSE
servers:
  - url: https://api.gamecomponents.io/v1
    description: Production server
  - url: https://api.staging.gamecomponents.io/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /tokens/{tokenId}:
    get:
      tags:
        - Tokens
      summary: Get token details
      description: |
        Retrieves detailed information about a specific token, including:
        - Decoded packed token ID fields (immutable metadata)
        - Current mutable state (game_over, completed_all_objectives)
        - Owner and player information
        - Optional score history

        The token ID encodes immutable metadata directly (game_id, minted_by, etc.)
        which is decoded and returned in the response.
      operationId: getToken
      parameters:
        - name: tokenId
          in: path
          required: true
          description: The unique token identifier (felt252 as decimal string)
          schema:
            type: string
            pattern: "^[0-9]+$"
            example: "1234567890123456789012345678901234567890"
        - name: includeScoreHistory
          in: query
          description: Include historical score entries
          schema:
            type: boolean
            default: false
        - name: scoreHistoryLimit
          in: query
          description: Maximum number of score history entries to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Token details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDetail"
              examples:
                activeGame:
                  summary: Active game token
                  value:
                    tokenId: "1234567890123456789012345678901234567890"
                    gameId: 1
                    game:
                      id: 1
                      name: "Loot Survivor"
                      contractAddress: "0x0123..."
                    owner: "0xabc123..."
                    playerName: "Player1"
                    currentScore: 4500
                    rank: 15
                    gameOver: false
                    completedAllObjectives: false
                    soulbound: false
                    mintedAt: "2024-01-15T10:30:00Z"
                    mintedBy:
                      id: 1
                      name: "Official Minter"
                    settingsId: 1
                    lifecycleStart: "2024-01-15T10:30:00Z"
                    lifecycleEnd: "2024-01-22T10:30:00Z"
                    objectivesCount: 5
                    hasContext: true
                    sequenceNumber: 12345
                completedGame:
                  summary: Completed game token
                  value:
                    tokenId: "9876543210987654321098765432109876543210"
                    gameId: 1
                    owner: "0xdef456..."
                    playerName: "Champion"
                    currentScore: 15000
                    rank: 1
                    gameOver: true
                    completedAllObjectives: true
                    soulbound: true
                    mintedAt: "2024-01-10T08:00:00Z"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Invalid token ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tokens:
    get:
      tags:
        - Tokens
      summary: List tokens with filtering
      description: |
        Query tokens with optional filters for game, owner, status, and score range.
        Results are paginated and can be sorted by various fields.
      operationId: listTokens
      parameters:
        - name: gameId
          in: query
          description: Filter by game ID
          schema:
            type: integer
        - name: owner
          in: query
          description: Filter by owner address
          schema:
            type: string
        - name: status
          in: query
          description: Filter by game status
          schema:
            type: string
            enum: [active, completed, all]
            default: all
        - name: minScore
          in: query
          description: Minimum score filter
          schema:
            type: integer
            minimum: 0
        - name: maxScore
          in: query
          description: Maximum score filter
          schema:
            type: integer
            minimum: 0
        - name: soulbound
          in: query
          description: Filter by soulbound status
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: orderBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [mintedAt, currentScore, lastUpdatedAt, completedAt]
            default: mintedAt
        - name: orderDir
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: List of tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenListResponse"
...
  /players/{address}/tokens:
    get:
      tags:
        - Players
      summary: Get player's tokens
      description: |
        Retrieves all tokens owned by a specific player address.
        Supports filtering by game and game status.
      operationId: getPlayerTokens
      parameters:
        - name: address
          in: path
          required: true
          description: Player wallet address (hex format with 0x prefix)
          schema:
            type: string
            pattern: "^0x[a-fA-F0-9]+$"
        - name: gameId
          in: query
          description: Filter by game ID
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by game status
          schema:
            type: string
            enum: [active, completed, all]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [mintedAt, currentScore, lastUpdatedAt, completedAt]
            default: lastUpdatedAt
        - name: orderDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Player's tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerTokensResponse"
        "400":
          description: Invalid address format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    TokenDetail:
      type: object
      required:
        - tokenId
        - gameId
        - owner
        - currentScore
        - gameOver
        - completedAllObjectives
        - soulbound
        - mintedAt
        - objectivesCount
        - hasContext
        - sequenceNumber
      properties:
        tokenId:
          type: string
          description: Unique token identifier (felt252 as decimal string)
        gameId:
          type: integer
          description: Game identifier (decoded from packed token ID)
        game:
          $ref: "#/components/schemas/GameSummary"
        owner:
          type: string
          description: Current owner address
        playerName:
          type: string
          nullable: true
          description: Player display name
        currentScore:
          type: integer
          description: Current game score
        rank:
          type: integer
          nullable: true
          description: Current leaderboard rank (null if not on leaderboard)
        gameOver:
          type: boolean
          description: Whether the game has ended
        completedAllObjectives:
          type: boolean
          description: Whether all objectives were completed
        completedAt:
          type: integer
          nullable: true
          description: Unix timestamp when objectives were completed
        soulbound:
          type: boolean
          description: Whether the token is non-transferable
        mintedAt:
          type: string
          format: date-time
          description: Token mint timestamp
        mintedBy:
          $ref: "#/components/schemas/MinterSummary"
        settingsId:
          type: integer
          description: Game settings ID used for this token
        lifecycleStart:
          type: string
          format: date-time
          nullable: true
          description: Game lifecycle start time
        lifecycleEnd:
          type: string
          format: date-time
          nullable: true
          description: Game lifecycle end time
        objectivesCount:
          type: integer
          description: Number of objectives for this game
        hasContext:
          type: boolean
          description: Whether this token has context data
        sequenceNumber:
          type: integer
          description: Unique sequence number within the game
        scoreHistory:
          type: array
          items:
            $ref: "#/components/schemas/ScoreEntry"
          description: Historical score entries (if requested)
        createdAtBlock:
          type: integer
          description: Block number when token was indexed
        lastUpdatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    TokenSummary:
      type: object
      required:
        - tokenId
        - gameId
        - owner
        - currentScore
        - gameOver
      properties:
        tokenId:
          type: string
        gameId:
          type: integer
        owner:
          type: string
        playerName:
          type: string
          nullable: true
        currentScore:
          type: integer
        gameOver:
          type: boolean
        soulbound:
          type: boolean
        mintedAt:
          type: string
          format: date-time
    ScoreEntry:
      type: object
      required:
        - score
        - blockNumber
        - blockTimestamp
      properties:
        score:
          type: integer
          description: Score value at this point
        blockNumber:
          type: integer
          description: Block number when score was recorded
        blockTimestamp:
          type: string
          format: date-time
          description: Timestamp when score was recorded
...
    GameSummary:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        contractAddress:
          type: string
...
    MinterSummary:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        contractAddress:
          type: string
...
    TokenListResponse:
      type: object
      required:
        - tokens
        - totalCount
        - pagination
      properties:
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/TokenSummary"
        totalCount:
          type: integer
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
...
    PlayerTokensResponse:
      type: object
      required:
        - address
        - tokens
        - totalCount
      properties:
        address:
          type: string
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/TokenSummary"
        totalCount:
          type: integer
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
...
    PaginationInfo:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true
          description: Cursor for next page (if using cursor pagination)
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_REQUEST
            - NOT_FOUND
            - VALIDATION_ERROR
            - RATE_LIMITED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        requestId:
          type: string
          description: Request ID for support reference
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for rate limit tracking and analytics
security:
  - ApiKeyAuth: []
  - {}  # Also allow unauthenticated access with stricter rate limits